requirements.txt
streamlit>=1.29.0
yt-dlp>=2023.7.6
moviepy==1.0.3  # Pinned to 1.0.3 for compatibility with moviepy.editor
static-ffmpeg>=2.13.0
openai>=1.3.0
openai-whisper>=20231117
torch>=2.2.0
python-dotenv>=1.0.0
rouge-score>=0.1.2
bert-score>=0.3.13
